Refactor nextValue()
In order to allow the operator to be changed, this commit adds a check
to see if the last pressed key (`lastValue`) is an operator. If it is,
and the current value (`value`) is an operator, it updates the operator
without calculating anything.

This commit also removes the redundant `newValue` boolean in favor of
using the similar functionality of checking if the last value was an
operator.